Initial Commit
Following items work:
- ACMEEnroll
-- playbook is fully functional
- ACMEVerify
-- Generate CSRs for all needed certs
-- Submit CSRs to ACME and receive challenges

Next big todo: Take ACME challenges and parse them out.  We should save the "register" output of the challenges then parse with external python.  We need to get a list of R53 domains we have, then check the challenges vs. the domains and add a "domain" arg to each challenge so we can actually run the R53 commands.